Added support for ouput paging in ghi-ls
This feature addition borrows *heavily* from pager.py of the AOSP
project 'repo'. This has been done out of the utmost respect for that
project and its authors. Their solution to output paging within a python
project is excellent and I would not suggest that I could do it better.